RegEx and Functional Updates
Made port number optional in RegEx.
Workaround to allow reading of selected data in TEXTAREA.
In the URL regex replaced // with \/. I think was was a typo to include the forward slash which needs to be escaped, I think.
Included the whitespace character int he URL regex to match URLs with spaces.
Updated RegEx to match on port numbers in IPs and domains as well.